# lib/api-zod/src/generated/types/penugasanUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sistem Koordinasi Bencana Bali API
 * OpenAPI spec version: 0.1.0
 */
import type { PenugasanUpdateStatus } from './penugasanUpdateStatus';

export interface PenugasanUpdate {
  status?: PenugasanUpdateStatus;
  /** @nullable */
  catatan?: string | null;
  /** @nullable */
  link_grup?: string | null;
}