# lib/api-zod/src/generated/types/transaksiInventarisInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sistem Koordinasi Bencana Bali API
 * OpenAPI spec version: 0.1.0
 */
import type { TransaksiInventarisInputJenis } from './transaksiInventarisInputJenis';

export interface TransaksiInventarisInput {
  jenis: TransaksiInventarisInputJenis;
  jumlah: number;
  keterangan: string;
  /** @nullable */
  laporan_id?: number | null;
}